[SofaGraphComponent] RENAME files according to class names
