+++ applications/sofa/gui/qt/RealGUI.cpp
@@ -116,7 +116,7 @@ using sofa::helper::system::FileMonitor;
 #include <SofaGraphComponent/SceneCheckerVisitor.h>
 using sofa::simulation::scenechecking::SceneCheckerVisitor;
 
-#include <SofaGraphComponent/SceneCheckAPIChanges.h>
+#include <SofaGraphComponent/SceneCheckAPIChange.h>
 using sofa::simulation::scenechecking::SceneCheckAPIChange;
 #include <SofaGraphComponent/SceneCheckMissingRequiredPlugin.h>
 using sofa::simulation::scenechecking::SceneCheckMissingRequiredPlugin;
