+++ modules/SofaGraphComponent/SceneCheck.h
